refactor : view패키지에 input, output view 이동
