+++ src/main/java/racegame/view/InputView.java
@@ -1,4 +1,4 @@
-package racegame;
+package racegame.view;
 
 import racingcar.Car;
 import utils.NameValidator;
