+++ src/main/java/racegame/view/OutputView.java
@@ -1,4 +1,4 @@
-package racegame;
+package racegame.view;
 
 import racingcar.Car;
